Move stream classes into uleb.h
