+++ tagging_tools/uleb.h
@@ -28,6 +28,66 @@
 #define ULEB_H
 
 #include <cstdint>
+#include <cstring>
+#include <fstream>
+#include <string>
+
+namespace policy_engine {
+
+class stream_reader_t {
+private:
+  std::ifstream is;
+  std::streamsize size;
+
+public:
+  stream_reader_t(const std::string& fname, std::ios::openmode mode=std::ios::binary) : is(std::ifstream(fname, mode)) {
+    is.ignore(std::numeric_limits<std::streamsize>::max());
+    size = is.gcount();
+    is.clear();
+    is.seekg(0, std::ios::beg);
+  }
+
+  template<class T> std::streamsize read(T* data, std::streamsize n) {
+    try {
+      std::streamsize b = n*sizeof(T)/sizeof(std::ofstream::char_type);
+      std::ofstream::char_type bytes[b];
+      std::streamsize r = is.read(bytes, b).gcount();
+      if (r == b)
+        std::memcpy(data, bytes, b);
+      return r*sizeof(std::ofstream::char_type)/sizeof(T);
+    } catch (const std::ios::failure& e) {
+      return 0;
+    }
+  }
+
+  bool read_byte(uint8_t& b) { return read(&b, 1) == 1; }
+
+  std::streamsize length() { return size; }
+  bool eof() { return is.tellg() >= size; }
+
+  explicit operator bool() const { return static_cast<bool>(is); }
+};
+
+class stream_writer_t {
+private:
+  std::ofstream os;
+
+public:
+  stream_writer_t(const std::string& fname, std::ios::openmode mode=std::ios::binary) : os(std::ofstream(fname, mode)) {}
+
+  template<class T> bool write(const T* data, std::size_t n) {
+    try {
+      os.write(reinterpret_cast<const std::ofstream::char_type*>(data), n*sizeof(T)/sizeof(std::ofstream::char_type));
+      return !os.fail();
+    } catch (const std::ios::failure& e) {
+      return false;
+    }
+  }
+
+  bool write_byte(uint8_t b) { return write(&b, 1); }
+
+  explicit operator bool() const { return static_cast<bool>(os); }
+};
 
 /**
  *  ULEB encoding functions.  These are generics that allow you to specify a reader type and
@@ -81,4 +141,6 @@ template <typename StreamType, typename ValueType> bool write_uleb(StreamType& s
   return true;
 }
 
-#endif
+} // namespace policy_engine
+
+#endif // ULEB_H
